+++ ProjectAlchemy.Core/Dtos/CreateProjectRequest.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using ProjectAlchemy.Core.Domain;
 
-namespace ProjectAlchemy.Web.Dtos;
+namespace ProjectAlchemy.Core.Dtos;
 
 public class CreateProjectRequest
 {
