+++ ProjectAlchemy.Core/Dtos/PartialIssue.cs
@@ -1,22 +1,21 @@
 using ProjectAlchemy.Core.Domain;
 using ProjectAlchemy.Core.Enums;
 
-namespace ProjectAlchemy.Web.Dtos;
+namespace ProjectAlchemy.Core.Dtos;
 
-public class IssueResponse
+public class PartialIssue
 {
     public required int Id { get; set; }
     public required string Name { get; set; }
-    public required string Description { get; set; }
     public required IssueType Type { get; set; }
     public required LaneView Lane { get; set; }
-    public static IssueResponse FromIssue(Issue issue)
+
+    public static PartialIssue FromIssue(Issue issue)
     {
-        return new IssueResponse()
+        return new PartialIssue()
         {
             Id = issue.Id,
             Name = issue.Name,
-            Description = issue.Description,
             Type = issue.Type,
             Lane = LaneView.FromLane(issue.Lane)
         };
