+++ ProjectAlchemy.Core/Interfaces/IMemberRepository.cs
@@ -1,4 +1,5 @@
 using ProjectAlchemy.Core.Domain;
+using ProjectAlchemy.Core.Dtos;
 
 namespace ProjectAlchemy.Core.Interfaces;
