+++ ProjectAlchemy.Persistence/Repositories/MemberRepository.cs
@@ -1,5 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using ProjectAlchemy.Core.Domain;
+using ProjectAlchemy.Core.Dtos;
 using ProjectAlchemy.Core.Interfaces;
 
 namespace ProjectAlchemy.Persistence.Repositories;
